transform/rtcalls: Optimize []byte to string conversions
This enables the optimization of functions such as "bytes.Equal()", which
is currently implemented as:

	func Equal(a, b []byte) bool { return string(a) == string(b) }

Which causes allocations and copies in TinyGo, but not in Big Go.

Since we can now prove that neither conversions will escape, we can convert
the calls to runtime.stringFromBytes() to an inlined version of it that doesn't
allocate or copies the original string.

Closes: #4045
Signed-off-by: L. Pereira <l.pereira@fastly.com>