[DRAFT] runtime: Rewrite hashmap data structure
This structure now deviates quite a bit from the one used in the Big Go
implementation.  It's inspired by swisstable, but with the mindset that
this is to be used in embedded systems; a description of the structure
can be found in hashmap.go.

Mainly, this structure tries its hardest to have a high load factor,
and performs linear probing to handle collisions.  The linear probing
is performed in a very efficient manner, by using fast byte scans,
rather than a set number of branches/bucket.  Only the top 8 bits of
the 32-bit hash is used, like previously, but the lower 3 bits are
used to index a linked list of buckets that contain up to 8 key/value
pairs.  The top 4 bits are stored in a per-bucket bitmap; keys are
only compared if this nibble matches.

Some tests are not passing yet; I'm still debugging this.  It's
pretty promising, though.

Some things that still need to be implemented:
  - Randomized iteration.
  - Improved memory allocation when sizeHint is passed.  Right now, each bucket
    is allocated individually rather than in a group of nodes.
  - Recycling of buckets when all its slots become empty.

[1] https://www.usenix.org/legacy/events/sec03/tech/full_papers/crosby/crosby.pdf

Signed-off-by: L. Pereira <l.pereira@fastly.com>